add inheritance structure
